backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
pymysql
langchain
chromadb
langchain-community
langchain-huggingface
langchain-openai
sentence-transformers
